+++ ultimate/testData/spring/core/gutter/importConfigClasses/baz/Config2.kt
@@ -1,0 +1,10 @@
+package baz
+
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
+import foo.FooBean
+import bar.BarBean
+
+@ComponentScan(basePackageClasses = arrayOf(FooBean::class, BarBean::class))
+@Configuration
+open class Config2